requirements.txt
numpy==1.21.5
Pillow==9.2.0
Pillow==10.1.0
PyAutoGUI==0.9.54
pytesseract==0.3.10
